+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 - #137 Adds key binding editor
 - #151 Moves next undo/redo actions to separate bar, adds setting to toggle showing it
 - Adds mouse tooltips to status, yank register, next undo/redo action
+- Improves resiliency of loading and saving config and history
 - Changes recycle/delete to not push useless undo actions
 - Fixes copy names of folders to not use file extension logic
 - Fixes input text selection being reset to start when navigating search history with search strings of the same length